pyproject.toml
[project]
name = "psi-agenda"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tzdata (>=2025.2,<2026.0)"
]

[tool.poetry]
packages = [{include = "psi_agenda", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.9.9"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
httpx = "^0.28.1"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.ruff.lint]
preview= true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
pre_test = 'task lint'
run = 'uvicorn src.psi_agenda.app:app --host 0.0.0.0 --port 8000 --reload'
test = 'pytest -s --cov=src.psi_agenda -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
